Made TWI display work
* Added floating point number support
* Changed float back to double
* Reverted to old double to string converter

+++ code/XMEGA_A3BU_XPLAINED_DEMO1/XMEGA_A3BU_XPLAINED_DEMO1/src/adc_sensors.c
@@ -221,7 +221,7 @@ int16_t ntc_get_raw_value(void)
 int8_t ntc_get_temperature(void)
 {
 	int8_t retval = 0;
-	float ntc_sample = ntc_sensor_sample;
+	double ntc_sample = ntc_sensor_sample;
 	if (ntc_sensor_sample > 697) {
 		retval = (int8_t)((-0.0295 * ntc_sample) + 40.5);
 	} if (ntc_sensor_sample > 420) {
